err and diff should be int16. avr-gcc handles multiplication of int32 x int16 faster than int32 x int32 and these values were anyway restricted to int16

+++ src/ArduinoPID.cpp
@@ -98,14 +98,14 @@ int16_t ArduinoPID::compute(int16_t setpoint, int16_t measurement){
     }
 	
     //wrap-safe subtraction to avoid issues with int16_t overflow
-	int32_t err =  (int16_t)(uint16_t(setpoint) - uint16_t(measurement));
+	int16_t err =  (int16_t)(uint16_t(setpoint) - uint16_t(measurement));
 
 	//Calculating Proportional term:
 	int64_t output = pGain * err;
 
 	//Calculating Derivative term:
 	if (dGain != 0){
-        int32_t diff = (int16_t)(uint16_t(measurement) - uint16_t(lastMeasurement));
+        int16_t diff = (int16_t)(uint16_t(measurement) - uint16_t(lastMeasurement));
         lastMeasurement = measurement;
 		if (derivativeFiltering == NO_FILTERING){
 			output -= dGain * diff;
